Enable e2e
- enable e2e test on github action pipeline
